--- src/main/java/Command.java
@@ -1,5 +1,0 @@
-public abstract class Command {
-
-    public abstract void handleCommand(Ui ui);
-    public abstract boolean isExit();
-}
