+++ src/main/java/DukeException/DukeException.java
@@ -1,0 +1,7 @@
+package DukeException;
+class DukeException extends Exception {
+    public DukeException(String message) {
+        super(message);
+    }
+}
+
