+++ src/main/java/DukeException/EmptyDescriptionException.java
@@ -1,0 +1,7 @@
+package DukeException;
+
+public class EmptyDescriptionException extends DukeException {
+    public EmptyDescriptionException() {
+        super("☹ OOPS!!! The description of your task cannot be empty.");
+    }
+}
