+++ src/main/java/DukeException/InvalidInputException.java
@@ -1,0 +1,7 @@
+package DukeException;
+
+public class InvalidInputException extends DukeException {
+    public InvalidInputException() {
+        super("☹ OOPS!!! I'm sorry, but I don't know what that means :-(");
+    }
+}
