+++ src/main/java/Task/Deadline.java
@@ -1,6 +1,7 @@
+package Task;
 public class Deadline extends Task {
 
-    protected String by;
+    public String by;
 
     public Deadline(String description, String by) {
         super(description);
