+++ src/main/java/Task/Event.java
@@ -1,3 +1,4 @@
+package Task;
 public class Event extends Task {
 
     protected String from;
