+++ src/main/java/Task/Task.java
@@ -1,3 +1,4 @@
+package Task;
 public class Task {
     protected String description;
     protected boolean isDone;
